rename/rearrange execute_wait tests
